biomedical-rag-upgrade-scripts/requirements.txt
torch==2.6.0
transformers==4.36.2
faiss-cpu==1.13.2
numpy==1.26.4
tqdm==4.66.1
scikit-learn==1.4.0
nltk==3.8.1
datasets==2.16.1
sentencepiece==0.1.99
protobuf==4.25.2
sentence-transformers==2.5.1
